Data.Runs

Documentation

newtype Runs a

Lists with runs of equal values replaced with (number of elements in the run, the repeated value)

Constructors

Runs

Fields

getRuns :: [(Int, a)]

toRuns :: Eq a => [a] -> Runs a

Convert a list to Runs, one half of an isomorphism:

 forall a. Eq a => Iso' [a] (Runs a)