Lib

Description

Want: Efficient (I think we can get near constant-time) conversion between a point (Integer, Integer) and its "spiral number": Natural

Documentation

newtype Pt

A point in Z^2

Constructors

Pt

Fields

getPr :: (Integer, Integer)

unSpiral :: Natural -> Pt

undefined

Convert a Natural number to its point on the spiral

toSpiral :: Pt -> Natural

undefined

Convert a point to its position on the spiral

stepSpiral :: Pt -> Pt

undefined

Get the next position on the spiral

spiralExample :: ()

Spiral example

 (20, -2,  2) (21, -1,  2) (22, 0,  2) (23, 1,  3) (24, 2,  2) (25, 3,  2)
 (19, -2,  1) ( 6, -1,  1) ( 7, 0,  1) ( 8, 1,  1) ( 9, 2,  1) (26, 3,  1)
 (18, -2,  0) ( 5, -1,  0) ( 0, 0,  0) ( 1, 1,  0) (10, 2,  0) (27, 3,  0)
 (17, -2, -1) ( 4, -1, -1) ( 3, 0, -1) ( 2, 1, -1) (11, 2, -1) (28, 3, -1)
 (16, -2, -2) (15, -1, -2) (14, 0, -2) (13, 1, -2) (12, 2, -2) (29, 3, -2)
 (35, -2, -3) (34, -1, -3) (33, 0, -3) (32, 1, -3) (31, 2, -3) (30, 3, -3)

unwrappedSpiralExample :: ()

Unwrapped spiral example:

 (20, -2,  2)
 (19, -2,  1)
 (18, -2,  0)
 (17, -2, -1)
 (16, -2, -2)
 (21, -1,  2)
 ( 6, -1,  1)
 ( 5, -1,  0)
 ( 4, -1, -1)
 (15, -1, -2)
 (22, 0,  2)
 ( 7, 0,  1)
 ( 0, 0,  0)
 ( 3, 0, -1)
 (14, 0, -2)
 (23, 1,  3)
 ( 8, 1,  1)
 ( 1, 1,  0)
 ( 2, 1, -1)
 (13, 1, -2)
 (24, 2,  2)
 ( 9, 2,  1)
 (10, 2,  0)
 (11, 2, -1)
 (12, 2, -2)
 (25, 3,  2)
 (26, 3,  1)
 (27, 3,  0)
 (28, 3, -1)
 (29, 3, -2)

examples :: [Pt]

Point examples

examples2 :: [Pt]

ptSimpleAlgRelations :: ()

Point simple algebraic relations

 [begin, end] (begin, diff)
 mult: val

 [0..0] (0, 0)
 2: 1
 [0..0] (0, 0)
 3: -1
 [0..1] (0, 1)
 4: 2
 [1..-1] (1, -2)
 5: -2
 [-1..2] (-1, 3)
 6: 3
 [2..-2] (2, -4)
 7: -3
 [-2..3] (-2, 5)
 9: 4
 [3..-3] (3, -6)
 9: -4
 [-3..4] (-3, 7)
 10: 5
 [4..-4] (4, -8)

simpleAlgBegins :: [(Integer, Integer)]

From ptSimpleAlgRelations

simpleAlgDiffs :: [Integer]

simpleAlgMults :: [Integer]

simpleAlgVals :: ()

ptSimpleAlgRelations values:

  1 -1
  2 -2
  3 -3
  4 -4
  5

pairList :: a -> a -> [a]

zipWithTail :: (a -> a -> b) -> [a] -> [b]

zipWith on the tail of a list

 λ> zipWithTail pairList [1..5]
 [[2,1],[3,2],[4,3],[5,4]]

zipTail :: [a] -> [(a, a)]

zip on the tail of a list

 λ> zipTail [1..5]
 [(2,1),(3,2),(4,3),(5,4)]

diffs :: Num a => [a] -> [a]

Successive (first) differences

 λ> diffs [1..5]
 [1,1,1,1]